+++ spec/integration/index/update_spec.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+RSpec.describe '`spandx index update` command', type: :cli do
+  it 'executes `spandx index help update` command successfully' do
+    output = `spandx index help update`
+    expected_output = <<~OUT
+      Usage:
+        spandx index update
+
+      Options:
+        -h, [--help], [--no-help]  # Display usage information
+
+      Command description...
+    OUT
+
+    expect(output).to eq(expected_output)
+  end
+end